Cast type when loading masks

+++ agent/multi_task_single_obj_multi_model.py
@@ -245,7 +245,7 @@ class MultiTaskSingleObjectiveMultiModelAgent(MultiTaskSingleObjectiveSingleMode
         try:
             with open(os.path.join(path, 'masks'), 'r') as f:
                 self.finalmodel_mask = json.load(f)
-            self.finalmodel_mask = torch.tensor(self.finalmodel_mask)
+            self.finalmodel_mask = torch.tensor(self.finalmodel_mask, dtype=torch.uint8)
             self.finalmodel = [self.submodel(masks, t) for t, masks in enumerate(self.finalmodel_mask)]
             self.finalmodel = [nn.DataParallel(model).to(self.device) for model in self.finalmodel]
